backend/text-to-speech/requirements.txt
dotenv
modal
pydantic
fastapi
chatterbox-tts
torchaudio
